+++ Backend/src/db-connect.ts
@@ -1,0 +1,11 @@
+import postgres from 'postgres';
+
+const sql = postgres({
+  host: 'localhost',
+  port: 5432,
+  database: process.env.DB_NAME,
+  username: process.env.DB_USERNAME,
+  password: process.env.DB_PASSWORD,
+});
+
+export default sql;