+++ Frontend/src/hooks.server.ts
@@ -4,7 +4,7 @@ export const handle: Handle = async ({event, resolve}) => {
   if (event.url.pathname.startsWith('/api')) {
     const response = await fetch(`http://localhost:3000${event.url.pathname}`, event.request)
       .then((res) => res.json())
-      .catch(() => ({message: 'An error occurred. Please try again later.'}));
+      .catch(() => ({message: 'An error occurred. Please try again later.', type: 'alert'}));
 
     return new Response(JSON.stringify(response), {headers: {'Content-Type': 'application/json'}});
   }
@@ -12,4 +12,6 @@ export const handle: Handle = async ({event, resolve}) => {
   return resolve(event);
 };
 
-export const handleError: HandleServerError = () => ({message: 'An error occurred. Please try again later.'});
+export const handleError: HandleServerError = () => {
+  return {message: 'An error occurred. Please try again later.', type: 'alert'};
+};