+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2024-10-06
+
+### Added
+
+- Add ability to register to waitlist
+- Connect backend to database (postgres.js)
+
+### Changed
+
+- Move validation to server-side
+- Move database name and username to an environment variable
+- Fix small mistakes in frontend
+
 ## [0.0.4] - 2024-10-03
 
 ### Added